move PhotoPreviewModal out of Observation folder
