+++ src/frontend/Navigation/Stack/AppScreens.tsx
@@ -56,7 +56,7 @@ import {ObservationFields} from '../../screens/ObservationFields';
 import {LIGHT_GREY, WHITE} from '../../lib/styles';
 import {LanguageSettings} from '../../screens/Settings/AppSettings/LanguageSettings';
 import {BottomSheetModalProvider} from '@gorhom/bottom-sheet';
-import {PhotoPreviewModal} from '../../screens/Observation/PhotoPreviewModal.tsx';
+import {PhotoPreviewModal} from '../../screens/PhotoPreviewModal.tsx';
 import {CustomHeaderLeft} from '../../sharedComponents/CustomHeaderLeft.tsx';
 
 export const TAB_BAR_HEIGHT = 70;
